requirements.txt
asgiref==3.8.1
Django==5.1.2
django-redis==5.4.0
djangorestframework==3.15.2
gunicorn==20.1.0
numpy==2.1.2
pandas==2.2.3
pre-commit==4.1.0
psycopg==3.2.4
python-dotenv==1.0.1
ratelimit==2.2.1
redis==5.2.1
requests==2.32.3
ruff==0.9.6
yfinance==0.2.54
